+++ python/cudf/cudf/core/column/string.py
@@ -4421,9 +4421,8 @@ class StringColumn(column.ColumnBase):
 
         if len(children) == 0 and size != 0:
             # all nulls-column:
-            offsets = cudf.core.column.as_column(
-                cupy.zeros(size + 1, dtype="int32")
-            )
+            offsets = column.full(size + 1, 0, dtype="int32")
+
             chars = cudf.core.column.as_column([], dtype="int8")
             children = (offsets, chars)